:recycle: [wire,pkg/io/test] Move generic test wrapper from wire->io/test
- The wrapper for generic serializer tests on wire messages was defined
  in wire package.

- But, this should have been defined in a test helper package. So,
  moving it to the relevant test helper package io/test.

- Since, this internal tests cannot import io/test (because io/test
  imports wire), move the tests which use this helper function to
  external test package "wire_test".
